Improve test coverage of Setup() functions
